// apps/fusion/project/ws/app/src/lib/routes/skill-management/models/competency-model.ts
export interface IApprover {
  skill_group: string
  role: string
  role_name: string
}

export interface ICourseObj {
  sequence_id: number
  course_type: string
  content_id: string
  content_name: string
}
export interface IAuthResponse {
  roles: IApprover[]
}
// ...
export interface IAddSkillObj {
  skill_name: string
  skill_description: string
  skill_id?: string
  skill_level_description_Starter: string
  skill_level_description_Basic: string
  skill_level_description_Advanced: string
  skill_level_description_Expert: string
  // skill_group: string
  skill_category: string
  skill_manager: string
  skill_relevance: string
}